+++ database/migrations/2020_07_30_183919_create_documents_files_table.php
@@ -21,7 +21,7 @@ class CreateDocumentsFilesTable extends Migration
             $table->string('title');
             $table->string('file_path');
             $table->integer('price');
-            $table->unsignedInteger('situation_id');
+            $table->unsignedBigInteger('situation_id');
             $table->foreign('situation_id')->references('id')->on('situations');
             $table->timestamps();
         });
