+++ database/migrations/2020_07_30_184012_create_documents_keys_table.php
@@ -18,7 +18,7 @@ class CreateDocumentsKeysTable extends Migration
     {
         Schema::create('documents_keys', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('document_file_id');
+            $table->unsignedBigInteger('document_file_id');
             $table->string('title');
             $table->string('key');
             $table->timestamps();
